Add dynamic styling based on status and reset search functionality
A ButtonProps type was added to the StyledArticle component to handle dynamic styling based on the character’s status. The border color of the article now changes depending on whether the character is alive or not. Also, a reset search functionality was incorporated in CharactersPage, which allows the user to reset search inputs to their initial state.