Add new character creation feature
The update introduces a new feature that allows users to create new characters. The "New" link was added to the main navigation component, while the filter feature for the character page was adjusted to incorporate the new creation functionality. A new model type, CharacterDTOType, was also introduced to represent the new character being created.