Add NewCharacterForm component
This commit introduces a new component, NewCharacterForm, to keep our code DRY. The aforementioned form was previously included directly in the NewCharacterPage component. Now, the form is separated and imported into NewCharacterPage. This refactoring can enhance the readability and maintainability of the codebase. The useNavigate function from react-router-dom has also been added to navigate to the corresponding character's page upon creation.