Add optional link to character card
This commit makes it optional for a character card to be clickable and navigate to a detail page. This feature is introduced by adding a new 'isLinked' prop to the CharacterCard component. Now, only when 'isLinked' is true, the CharacterCard will be clickable and navigate to the detail page when clicked.